+++ cpp/src/strings/nvcategory_util.cpp
@@ -8,6 +8,7 @@
 #include <nvstrings/NVStrings.h>
 #include <rmm/rmm.h>
 #include <utilities/error_utils.hpp>
+#include <utilities/legacy/error_utils.hpp>
 #include <cudf/binaryop.hpp>
 #include <cudf/legacy/copying.hpp>