+++ CHANGELOG.md
@@ -46,6 +46,8 @@
 - PR #3135 Add nvtx utilities to cudf::nvtx namespace
 - PR #3021 Java host side concat of serialized buffers
 - PR #3138 Movey unary files to legacy
+- PR #3175 Set cmake cuda version variables
+- PR #3171 Move deprecated error macros to legacy
 - PR #3165 Java device memory size for string category
 
 ## Bug Fixes
@@ -152,6 +154,7 @@
 - PR #2788 Guide to Python UDFs
 - PR #2919 Change java API to use operators in groupby namespace
 - PR #2909 CSV Reader: Avoid row offsets host vector default init
+- PR #2834 DataFrame supports setting columns via attribute syntax `df.x = col`
 
 ## Bug Fixes